+++ Bapad/Bapad/x64/Debug/Bapad.log
@@ -1,2 +1,1 @@
-﻿  Bapad.cpp
-LINK : fatal error LNK1104: cannot open file 'C:\Users\90314\Documents\GitHub\Bapad\Bapad\x64\Debug\TextView.lib'
+﻿
