--- Bapad/Bapad/x64/Debug/Bapad.tlog/link.command.1.tlog
@@ -1,1 +1,0 @@
-ÿþ
