+++ Bapad/TextView/TextDocument.h
@@ -14,18 +14,11 @@ public:
 
     bool clear();
   
-    bool  offset_to_line(ULONG fileoffset, ULONG* lineno, ULONG* offset);
-    bool  getlineinfo(ULONG lineno, ULONG* fileoff, ULONG* length);
-
-    size_t getline(size_t lineno, wchar_t* buf, size_t len, ULONG* fileoff = 0);
-    LONG getline(ULONG lineno, ULONG offset, wchar_t* buf, size_t len, ULONG* fileoff = 0);
-    ULONG getdata(ULONG offset, wchar_t* buf, size_t len);
+    size_t getline(size_t lineno, wchar_t* buf, size_t len);
 
     size_t getLinecount();
 
     size_t getLongestline(int tabwidth);
-
-    const size_t getDocLength() const;
     
 private:
     bool init_linebuffer();
