+++ Bapad/TextView/TextView.h
@@ -1,13 +1,8 @@
 #pragma once
 #include "TextDocument.h"
-#include "FontStruct.h"
-#include "TextViewWin32.h"
 
-//#define LONGEST_LINE 0x100
+#define LONGEST_LINE 0x100
 
-const int	LONGEST_LINE = 0x100;
-const int	TEXTBUFSIZE	= 32;
-const int	MAX_FONTS = 32;
 
 class TextView
 {
@@ -16,45 +11,22 @@ public:
 	TextView(HWND hwnd);
 	~TextView();
 
-	//滚动等操作后重新绘制
 	LONG OnPaint();
 	LONG OnSetFont(HFONT hFont);
 	LONG OnSize(UINT nFlags, int width, int height);
+
 	LONG OnVScroll(UINT nSBCode, UINT nPos);
 	LONG OnHScroll(UINT nSBCode, UINT nPos);
 	LONG OnMouseWheel(int nDelta);
 
-	//鼠标
-	LONG OnMouseActivate(HWND hwndTop, UINT nHitTest, UINT nMessage);
-	LONG OnLButtonDown(UINT nFlags, int x, int y);
-	LONG OnLButtonUp(UINT nFlags, int x, int y);
-	LONG OnMouseMove(UINT nFlags, int x, int y);
-
-
-	LONG OnSetFocus(HWND hwndOld);
-	LONG OnKillFocus(HWND hwndNew);
-
-
 	LONG OpenFile(WCHAR* szFileName);
 	LONG ClearFile();
-
-	LONG AddFont(HFONT);
-	LONG SetFont(HFONT, int idx);
-	LONG SetLineSpacing(int nAbove, int nBelow);
-
 
 private:
 
 	void PaintLine(HDC hdc, ULONG line);
-
 	void TabbedExtTextOut(HDC hdc, RECT* rect, WCHAR* buf, size_t len);
-
-
 	void RefreshWindow();
-	LONG InvalidateRange(ULONG nStart, ULONG nFinish);
-
-	int	 TabWidth();
-
 
 	COLORREF GetColour(UINT idx);
 
@@ -63,15 +35,9 @@ private:
 	bool    PinToBottomCorner();
 	void	Scroll(int dx, int dy);
 
-
-
-
-	BOOL  MouseCoordToFilePos(int x, int y, ULONG* pnLineNo, ULONG* pnCharOffset, ULONG* pnFileOffset, int* px);
-
 	HWND	hWnd;
 
 	// Font-related data	
-	FONT	fontAttr[MAX_FONTS];
 	HFONT	font;
 	int		fontWidth;
 	int		fontHeight;
@@ -86,21 +52,41 @@ private:
 	int		windowLines;
 	int		windowColumns;
 
-	// Display related data
-	int		tabWidthchars;
-	ULONG	selectionStart;
-	ULONG	selectionEnd;
-	ULONG	cursorOffset;
-
-
 	// File-related data
 	ULONG	lineCount;
 
-
-	// Runtime related data
-	bool	mouseDown;
-
 	//should use smart pointer?
 	//TextDocument* m_pTextDoc;
-	std::unique_ptr<TextDocument> pTextDoc;
+	std::unique_ptr<TextDocument> textDoc;
 };
+
+//	TextView API declared here
+//
+ATOM RegisterTextView(HINSTANCE hInstance);
+HWND CreateTextView(HWND hwndParent);
+
+//
+//	TextView Window Messages defined here
+//
+#define TXM_BASE         (WM_USER)
+#define TXM_OPENFILE     (TXM_BASE + 0)
+#define TXM_CLEAR        (TXM_BASE + 1)
+
+//
+//	TextView Message Macros defined here
+//
+#define TextView_OpenFile(hwndTV, szFile)	SendMessage((hwndTV), TXM_OPENFILE, 0, (LPARAM)(szFile))
+#define TextView_Clear(hwndTV)				SendMessage((hwndTV), TXM_CLEAR, 0, 0)
+
+//
+//	TextView Macros defined here
+//
+#define TEXTVIEW_CLASS L"TextView32"
+
+//
+//	TextView colours
+//
+#define TXC_BACKGROUND		0			// normal background colour
+#define TXC_FOREGROUND		1			// normal foreground colour
+
+#define TXC_MAX_COLOURS		2			// keep this updated!
