+++ Bapad/TextView/x64/Debug/TextView.log
@@ -1,20 +1,1 @@
-﻿  TextView.cpp
-C:\Users\90314\Documents\GitHub\Bapad\Bapad\TextView\TextView.cpp(96,39): warning C4244: 'argument': conversion from 'WPARAM' to 'UINT', possible loss of data
-C:\Users\90314\Documents\GitHub\Bapad\Bapad\TextView\TextView.cpp(99,37): warning C4244: 'argument': conversion from 'WPARAM' to 'UINT', possible loss of data
-C:\Users\90314\Documents\GitHub\Bapad\Bapad\TextView\TextView.cpp(102,37): warning C4244: 'argument': conversion from 'WPARAM' to 'UINT', possible loss of data
-  TextViewFile.cpp
-C:\Users\90314\Documents\GitHub\Bapad\Bapad\TextView\TextViewFile.cpp(14,41): warning C4267: '=': conversion from 'size_t' to 'ULONG', possible loss of data
-C:\Users\90314\Documents\GitHub\Bapad\Bapad\TextView\TextViewFile.cpp(15,44): warning C4267: '=': conversion from 'size_t' to 'int', possible loss of data
-C:\Users\90314\Documents\GitHub\Bapad\Bapad\TextView\TextViewFile.cpp(35,40): warning C4267: '=': conversion from 'size_t' to 'ULONG', possible loss of data
-C:\Users\90314\Documents\GitHub\Bapad\Bapad\TextView\TextViewFile.cpp(36,43): warning C4267: '=': conversion from 'size_t' to 'int', possible loss of data
-  TextViewFont.cpp
-  TextViewMouse.cpp
-C:\Users\90314\Documents\GitHub\Bapad\Bapad\TextView\TextViewMouse.cpp(149,37): warning C4267: '=': conversion from 'size_t' to 'ULONG', possible loss of data
-C:\Users\90314\Documents\GitHub\Bapad\Bapad\TextView\TextViewMouse.cpp(164,6): error C2660: 'TextDocument::getline': function does not take 5 arguments
-C:\Users\90314\Documents\GitHub\Bapad\Bapad\TextView\TextDocument.h(17,12): message : see declaration of 'TextDocument::getline'
-C:\Users\90314\Documents\GitHub\Bapad\Bapad\TextView\TextViewMouse.cpp(176,15): error C3861: 'TextWidth': identifier not found
-C:\Users\90314\Documents\GitHub\Bapad\Bapad\TextView\TextViewMouse.cpp(195,13): error C3861: 'TextWidth': identifier not found
-  TextViewPaint.cpp
-C:\Users\90314\Documents\GitHub\Bapad\Bapad\TextView\TextViewPaint.cpp(91,82): warning C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data
-  TextViewScoll.cpp
-  Generating Code...
+﻿
