--- Bapad/TextView/x64/Debug/TextView.tlog/TextView.lastbuildstate
@@ -1,2 +1,0 @@
-PlatformToolSet=v142:VCToolArchitecture=Native32Bit:VCToolsVersion=14.27.29110:TargetPlatformVersion=10.0.18362.0:
-Debug|x64|C:\Users\90314\Documents\GitHub\Bapad\Bapad\|
